[dep][trace] Move tracing from example to ice
- need to review the opentracing tutorial ....
