+++ example/github/Gopkg.toml
@@ -48,14 +48,6 @@ ignored = [
   version = "1.1.0"
 
 [[constraint]]
-  name = "github.com/opentracing/opentracing-go"
-  version = "1.0.2"
-
-[[constraint]]
-  name = "github.com/uber/jaeger-client-go"
-  version = "2.11.2"
-
-[[constraint]]
   branch = "master"
   name = "golang.org/x/oauth2"
