+++ example/github/cmd/icehubd/main.go
@@ -2,16 +2,11 @@ package main
 
 import (
 	"fmt"
-	"io"
 	"os"
 	"runtime"
 	"strings"
-	"time"
 
-	"github.com/opentracing/opentracing-go"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	jgconfig "github.com/uber/jaeger-client-go/config"
 	"google.golang.org/grpc"
 
 	icli "github.com/at15/go.ice/ice/cli"
@@ -47,10 +42,6 @@ var log = logutil.Registry
 
 // global configuration instance
 var cfg common.ServerConfig
-
-// TODO: might need a registry of application instead of scatter variables around in main
-var tracer opentracing.Tracer
-var closer io.Closer
 
 // TODO: just here for testing out the log command, though it might possible to make it like db command to be part
 // of go.ice's built in command for managing common config
@@ -127,33 +118,7 @@ func mustLoadConfig() {
 	}
 }
 
-// FIXME: hacky function to play with tracing libraries
-// https://github.com/jaegertracing/jaeger/blob/master/examples/hotrod/pkg/tracing/init.go#L31
-func configTracer() error {
-	tcfg := jgconfig.Configuration{
-		Sampler: &jgconfig.SamplerConfig{
-			Type:  "const",
-			Param: 1,
-		},
-		Reporter: &jgconfig.ReporterConfig{
-			LogSpans: false, // TODO: when true, enables LoggingReporter that runs in parallel with the main reporter
-			// and logs all submitted spans. Main Configuration.Logger must be initialized in the code
-			// for this option to have any effect.
-			BufferFlushInterval: 1 * time.Second,
-			LocalAgentHostPort:  "localhost:6831",
-		},
-	}
-	// TODO: a better way to use gommon/log, current tree level hierarchy may not be enough ...
-	// TODO: the jaeger.Logger interface is so strange, Error(string) instead of Error(string, args ...interface{})
-	// jgconfig.Logger(log)
-	// TODO: Observer can be registered with the Tracer to receive notifications about new Spans.
-	var err error
-	tracer, closer, err = tcfg.New("service-a")
-	if err != nil {
-		return errors.Wrap(err, "can't create jaeger tracer")
-	}
-	return nil
-}
+
 
 func main() {
 	cli = icli.New(
@@ -162,7 +127,7 @@ func main() {
 		icli.Version(buildInfo),
 		icli.LogRegistry(log))
 	root := cli.Command()
-	dbc := idbcmd.NewCommand(func() (icfg.DatabaseManagerConfig, error) {
+	dbc := idbcmd.New(func() (icfg.DatabaseManagerConfig, error) {
 		if err := cli.LoadConfigTo(&cfg); err != nil {
 			return cfg.DatabaseManager, err
 		}
